requirements.txt
astroid==2.3.3
autopep8==1.5.1
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
colorama==0.4.3
configparser==5.0.0
DateTime==4.3
fake-useragent==0.1.11
future==0.18.2
Google-Search-API==1.1.14
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
multidict==4.7.5
mysql-connector==2.2.9
psycopg2==2.8.5
psycopg2-binary==2.8.5
py2exe==0.9.2.2
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.5.0
pyint==1.1.0
pylint==2.4.4
pyTelegramBotAPI==3.6.7
pytz==2019.3
PyYAML==5.3.1
requests==2.7.0
rsa==4.0
scapy==2.4.3
selenium==2.53.6
six==1.14.0
soupsieve==2.0
Telethon==1.11.3
Unidecode==1.1.1
urllib3==1.25.8
vcrpy==4.0.2
wrapt==1.11.2
XlsxWriter==1.2.8
yarl==1.4.2
zope.interface==5.1.0
